Move to .graphql files
