+++ package.json
@@ -79,6 +79,7 @@
     "@storybook/addon-storyshots": "3.2.17",
     "@storybook/react": "3.2.17",
     "babel-eslint": "8.0.3",
+    "babel-plugin-inline-import-graphql-ast": "2.0.4",
     "cypress": "1.1.4",
     "enzyme": "3.2.0",
     "eslint": "4.13.0",
@@ -105,7 +106,7 @@
     "prepush": "yarn run test --onlyChanged",
     "analyze": "cross-env ANALYZE=1 next build src",
     "build": "next build src && node ./src/lib/defaultLang",
-    "compile-queries": "persistgraphql src/graphql/ --js --extension=js --add_typename",
+    "compile-queries": "persistgraphql src/graphql/ --add_typename",
     "cypress:open": "cypress open",
     "dev": "nodemon --delay 1000ms -w .env -w package.json -w yarn.lock -w server.js server.js",
     "dev:inspect":
