+++ src/graphql/index.js
@@ -1,0 +1,24 @@
+export { default as ActivateNextBlockMutation } from './mutations/ActivateNextBlockMutation.graphql'
+export { default as AddConfusionTSMutation } from './mutations/AddConfusionTSMutation.graphql'
+export { default as AddFeedbackMutation } from './mutations/AddFeedbackMutation.graphql'
+export { default as AddResponseMutation } from './mutations/AddResponseMutation.graphql'
+export { default as CreateQuestionMutation } from './mutations/CreateQuestionMutation.graphql'
+export { default as CreateSessionMutation } from './mutations/CreateQuestionMutation.graphql'
+export { default as DeleteFeedbackMutation } from './mutations/DeleteFeedbackMutation.graphql'
+export { default as EndSessionMutation } from './mutations/EndSessionMutation.graphql'
+export { default as LoginMutation } from './mutations/LoginMutation.graphql'
+export { default as ModifyQuestionMutation } from './mutations/ModifyQuestionMutation.graphql'
+export { default as RegistrationMutation } from './mutations/RegistrationMutation.graphql'
+export { default as StartSessionMutation } from './mutations/StartSessionMutation.graphql'
+export {
+  default as UpdateSessionSettingsMutation,
+} from './mutations/UpdateSessionSettingsMutation.graphql'
+
+export { default as AccountSummaryQuery } from './queries/AccountSummaryQuery.graphql'
+export { default as JoinSessionQuery } from './queries/JoinSessionQuery.graphql'
+export { default as QuestionDetailsQuery } from './queries/QuestionDetailsQuery.graphql'
+export { default as QuestionListQuery } from './queries/QuestionListQuery.graphql'
+export { default as RunningSessionQuery } from './queries/RunningSessionQuery.graphql'
+export { default as SessionEvaluationQuery } from './queries/SessionEvaluationQuery.graphql'
+export { default as SessionListQuery } from './queries/SessionListQuery.graphql'
+export { default as TagListQuery } from './queries/TagListQuery.graphql'
