+++ src/graphql/mutations/CreateQuestionMutation.graphql
@@ -1,0 +1,60 @@
+mutation CreateQuestion(
+  $title: String!
+  $description: String!
+  $options: QuestionOptionsInput!
+  $solution: Question_SolutionInput
+  $type: Question_Type!
+  $tags: [ID!]!
+) {
+  createQuestion(
+    question: {
+      title: $title
+      description: $description
+      options: $options
+      solution: $solution
+      type: $type
+      tags: $tags
+    }
+  ) {
+    id
+    title
+    type
+    tags {
+      id
+      name
+    }
+    versions {
+      id
+      description
+      options {
+        SC {
+          choices {
+            correct
+            name
+          }
+          randomized
+        }
+        MC {
+          choices {
+            correct
+            name
+          }
+          randomized
+        }
+        FREE_RANGE {
+          restrictions {
+            min
+            max
+          }
+        }
+      }
+      solution {
+        SC
+        MC
+        FREE
+        FREE_RANGE
+      }
+      createdAt
+    }
+  }
+}
