+++ src/graphql/mutations/EndSessionMutation.graphql
@@ -1,0 +1,6 @@
+mutation EndSession($id: ID!) {
+  endSession(id: $id) {
+    id
+    status
+  }
+}
