+++ src/graphql/mutations/LoginMutation.graphql
@@ -1,0 +1,10 @@
+mutation Login($email: String!, $password: String!) {
+  login(email: $email, password: $password) {
+    id
+    email
+    shortname
+    runningSession {
+      id
+    }
+  }
+}
