+++ src/graphql/mutations/StartSessionMutation.graphql
@@ -1,0 +1,6 @@
+mutation StartSession($id: ID!) {
+  startSession(id: $id) {
+    id
+    status
+  }
+}
