+++ src/graphql/queries/AccountSummaryQuery.graphql
@@ -1,0 +1,9 @@
+query AccountSummary {
+  user {
+    id
+    shortname
+    runningSession {
+      id
+    }
+  }
+}
