+++ src/graphql/queries/JoinSessionQuery.graphql
@@ -1,0 +1,44 @@
+query JoinSession($shortname: String!) {
+  joinSession(shortname: $shortname) {
+    id
+    settings {
+      isFeedbackChannelActive
+      isFeedbackChannelPublic
+      isConfusionBarometerActive
+    }
+    activeQuestions {
+      id
+      instanceId
+      title
+      description
+      type
+      options {
+        FREE_RANGE {
+          restrictions {
+            min
+            max
+          }
+        }
+        SC {
+          choices {
+            correct
+            name
+          }
+          randomized
+        }
+        MC {
+          choices {
+            correct
+            name
+          }
+          randomized
+        }
+      }
+    }
+    feedbacks {
+      id
+      content
+      votes
+    }
+  }
+}
