+++ src/graphql/queries/QuestionDetailsQuery.graphql
@@ -1,0 +1,47 @@
+query QuestionDetails($id: ID!) {
+  question(id: $id) {
+    id
+    title
+    type
+    instances {
+      id
+      createdAt
+    }
+    tags {
+      id
+      name
+    }
+    versions {
+      id
+      description
+      options {
+        SC {
+          choices {
+            correct
+            name
+          }
+        }
+        MC {
+          choices {
+            correct
+            name
+          }
+        }
+        FREE_RANGE {
+          restrictions {
+            min
+            max
+          }
+        }
+      }
+      solution {
+        SC
+        MC
+        FREE
+        FREE_RANGE
+      }
+    }
+    createdAt
+    updatedAt
+  }
+}
