+++ src/graphql/queries/QuestionListQuery.graphql
@@ -1,0 +1,22 @@
+query QuestionList {
+  questions: allQuestions {
+    id
+    title
+    type
+    instances {
+      id
+      createdAt
+    }
+    tags {
+      id
+      name
+    }
+    versions {
+      id
+      description
+      createdAt
+    }
+    createdAt
+    updatedAt
+  }
+}
