+++ src/graphql/queries/SessionListQuery.graphql
@@ -1,0 +1,21 @@
+query SessionList {
+  sessions: allSessions {
+    id
+    name
+    status
+    blocks {
+      id
+      instances {
+        id
+        version
+        question {
+          id
+          title
+          type
+        }
+      }
+    }
+    createdAt
+    updatedAt
+  }
+}
