+++ src/lang/en.json
@@ -1,10 +1,12 @@
 {
-  "common.string.skip": "Skip",
-  "common.string.submit": "Submit",
+  "common.button.skip": "Skip",
+  "common.button.submit": "Submit",
   "common.string.settings": "Settings",
   "common.string.logout": "Logout",
-  "runningSession.confusionBarometer.string.title": "Confusion-Barometer",
+  "runningSession.confusion.title": "Confusion-Barometer",
+  "confusionSection.noData": "No data yet.",
   "teacher.evaluation.graph.showGraph": "Show Graph",
+  "teacher.evaluation.graph.noResults": "No Results Available",
   "teacher.evaluation.totalParticipants.label": "Total participants:",
   "teacher.evaluation.possibilities.choices": "Choices",
   "teacher.evaluation.possibilities.restrictions": "Restrictions",
@@ -13,13 +15,15 @@
   "teacher.createQuestion.previewLabel": "Audience Preview",
   "common.button.discard": "Discard",
   "common.button.save": "Save",
+  "teacher.editQuestion.type": "Question Type",
   "form.useCase.label": "Use case description",
   "form.button.submit": "Submit",
+  "teacher.sessionCreation.title": "Create Session",
   "common.button.start": "Start",
   "student.sidebar.activeQuestion": "Active Question",
   "student.sidebar.feedbackChannel": "Feedback-Channel",
   "teacher.questionPool.title": "Question Pool",
-  "teacher.sessionHistory.title": "Sessions",
+  "teacher.sessionList.title": "Session List",
   "teacher.runningSession.title": "Running Session",
   "teacher.createQuestion.contentInput.label": "Question",
   "teacher.createQuestion.contentInput.tooltip": "Enter the question you want to ask the audience.",
@@ -30,35 +34,46 @@
   "teacher.createQuestion.titleInput.tooltip":
     "Enter a short summarizing title for the question. This is only visible to you!",
   "questionPool.question.lastUsed": "Last used",
+  "questionList.string.noQuestions": "No questions available.",
+  "tagList.string.noTags": "No tags available.",
   "teacher.createQuestion.options.min": "Min",
   "teacher.createQuestion.options.max": "Max",
   "teacher.createQuestion.optionsFREE.label": "Input Restrictions",
   "teacher.createQuestion.optionsFREE.tooltip": "Choose the allowed format of incoming responses.",
   "teacher.createQuestion.selection": "Selection",
-  "common.button.submit": "Submit",
   "teacher.createQuestion.optionsSC.label": "Available Choices",
   "teacher.createQuestion.optionsSC.tooltip":
     "Add answering options the respondents can choose from.",
   "teacher.createQuestion.questionType.label": "Question Type",
   "teacher.createQuestion.questionType.tooltip":
     "Choose the type of question you would like to create.",
+  "teacher.sessionCreation.newBlock": "New Block",
+  "teacher.sessionCreation.emptyDropzoneInfo": "Drag & drop a question into the dropzone.",
   "common.string.speed": "Speed",
   "common.string.difficulty": "Difficulty",
   "joinSession.newFeedbackInput.label": "New Feedback",
   "common.form.submit": "Submit",
-  "joinSession.allQuestionsCompleted": "You have completed all active questions.",
-  "sessionHistory.string.createdOn": "Created on",
-  "sessionHistory.title.runningSession": "Running session",
-  "sessionHistory.string.noSessionRunning": "No session is currently running.",
-  "sessionHistory.title.remainingSessions": "Remaining sessions",
+  "student.questionArea.info.SC": "Please choose a single option below:",
+  "student.questionArea.info.MC": "Please choose one or multiple of the options below:",
+  "student.questionArea.info.FREE": "Please enter your response below:",
+  "student.questionArea.info.FREE_RANGE": "Please choose a number from the given range below:",
+  "student.questionArea.title": "Question",
+  "joinSession.questionArea.alreadyAnswered": "You have already answered all active questions.",
+  "sessionList.string.createdOn": "Created on",
+  "sessionList.string.noSessions": "No session was found.",
+  "sessionList.title.runningSession": "Running session",
+  "sessionList.string.noSessionRunning": "No session is currently running.",
+  "sessionList.title.plannedSessions": "Planned sessions",
+  "sessionList.title.completedSessions": "Completed sessions",
   "session.button.completed.content": "Copy",
   "session.button.created.content": "Start",
   "session.button.running.content": "Running",
   "runningSession.button.evaluation": "Evaluation",
   "joinSession.noEvaluationActive": "No evaluation active.",
+  "joinSession.noFeedbackChannel": "Feedback-Channel deactivated.",
   "joinSession.noSessionActive": "No session active.",
-  "questionPool.button.createSession": "Create Session",
   "questionPool.button.createQuestion": "Create Question",
+  "questionPool.button.createSession": "Create Session",
   "common.string.send": "Send",
   "user.login.title": "Login",
   "user.registration.title": "Registration",
