+++ src/pages/join.js
@@ -16,12 +16,12 @@ import { graphql } from 'react-apollo'
 import FeedbackArea from '../components/sessions/join/FeedbackArea'
 import QuestionArea from '../components/sessions/join/QuestionArea'
 import { pageWithIntl, withData, withFingerprint } from '../lib'
-import { JoinSessionQuery } from '../graphql/queries'
 import {
+  JoinSessionQuery,
   AddConfusionTSMutation,
   AddFeedbackMutation,
   AddResponseMutation,
-} from '../graphql/mutations'
+} from '../graphql'
 import { StudentLayout } from '../components/layouts'
 
 const propTypes = {
