+++ src/pages/user/login.js
@@ -8,7 +8,7 @@ import { graphql } from 'react-apollo'
 
 import { StaticLayout } from '../../components/layouts'
 import { LoginForm } from '../../components/forms'
-import { LoginMutation } from '../../graphql/mutations'
+import { LoginMutation } from '../../graphql'
 import { pageWithIntl, withData } from '../../lib'
 
 const propTypes = {
